Move node-specific common utils into separate directory
