+++ packages/common/src/index.ts
@@ -2,13 +2,13 @@ export * as check from './check'
 export * as util from './util'
 
 export * from './async'
-export * from './fs'
 export * from './util'
 export * from './tid'
 export * from './ipld'
-export * from './ipld-cbor'
-export * from './logger'
 export * from './types'
-export * from './streams'
 export * from './times'
 export * from './strings'
+export * from './node/fs'
+export * from './node/ipld-cbor'
+export * from './node/logger'
+export * from './node/streams'
