+++ packages/common/src/node/streams.ts
